lotusapp/requirements-test.txt
pytest==8.4.1
pytest-django==4.11.1
pytest-cov==6.2.1
factory-boy==3.3.3
